refactor(board-sets): match filenames to component names
